Feat: 디렉토리 구조 변경 (기본정보, 기능, 테마)
