--- src/components/form/AccountSelection/AccountSelection.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import AccountInput from './AccountInput';
-
-const AccountSelection = () => {
-  return (
-    <div>
-      <AccountInput />
-    </div>
-  );
-};
-
-export default AccountSelection;
