--- src/components/form/ContactSelection/ContactSelection.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ContactInput from './ContactInput';
-
-const ContactSelection = () => {
-  return (
-    <div>
-      <ContactInput />
-    </div>
-  );
-};
-
-export default ContactSelection;
