+++ src/components/form/Feature/AccountFeature/AccountFeature.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import AccountInput from './AccountInput';
+
+const AccountFeature = () => {
+  return (
+    <div>
+      <AccountInput />
+    </div>
+  );
+};
+
+export default AccountFeature;
