+++ src/components/form/Feature/ContactFeature/ContactFeature.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import ContactInput from './ContactInput';
+
+const ContactFeature = () => {
+  return (
+    <div>
+      <ContactInput />
+    </div>
+  );
+};
+
+export default ContactFeature;
