+++ src/components/form/Feature/InformationFeature/InformationInput.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const InformationFeature = () => {
+  return <div>InformationInput</div>;
+};
+
+export default InformationFeature;
