+++ src/components/form/Feature/LocationFeature/LocationFeature.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const LocationFeature = () => {
+  return <div>LocationFeature</div>;
+};
+
+export default LocationFeature;
