--- src/components/form/InformationInput/InformationInput.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const InformationInput = () => {
-  return <div>InformationInput</div>;
-};
-
-export default InformationInput;
