--- src/components/form/LocationSelection/LocationSelection.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const LocationSelection = () => {
-  return <div>LocationSelection</div>;
-};
-
-export default LocationSelection;
