--- src/components/form/MessageSelection/MessageSelection.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const MessageSelection = () => {
-  return <div>MessageSelection</div>;
-};
-
-export default MessageSelection;
