+++ src/components/form/Theme/ColorFeature/ColorFeature.tsx
@@ -1,0 +1,5 @@
+const ColorFeature = () => {
+  return <div>ColorFeature</div>;
+};
+
+export default ColorFeature;
